+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added Aphrodite Proximity Indicator
+- Added Sprint Hitbox
+- Added Chronos pause
+
 ### Changed
 
 - Ported to h2m format